add Maison + matiere API

+++ src/main/java/Serdaigle/MIAGIE/repository/MatiereRepository.java
@@ -7,7 +7,7 @@ import Serdaigle.MIAGIE.model.Matiere;
 import java.util.Optional;
 
 @Repository
-public interface MatiereRepository extends JpaRepository<Matiere, Long> {
+public interface MatiereRepository extends JpaRepository<Matiere, Integer> {
 
     Matiere findByNomMatiere(String nomMatiere);
 }
